update SPIClass::setClockDivider to use uint16_t arg to allow for /256 divider

+++ libraries/SPI/src/SPI.cpp
@@ -47,7 +47,7 @@ void SPIClass::end()
     SPI_DeInit(this->config->register_base);
 }
 
-void SPIClass::setClockDivider(const uint8_t divider)
+void SPIClass::setClockDivider(const uint16_t divider)
 {
     en_spi_clk_div_t ddl_divider;
     switch(divider)
